# Clean up jvm double tensor broadcast