This PR adds a minor update to how broadcasting works in the JVMDoubleTensor.  It improves things a bit when the broadcast is large (e.g. 20x200 * 20x20x200).

Before this PR:

```
Benchmark                                            (dims)  (impl)  (operation)   Mode  Cnt    Score   Error  Units
BroadcastBinaryDoubleOperations.benchmark  20x200,20x20x200     JVM        TIMES  thrpt   15  177.820 ± 1.997  ops/s
BroadcastBinaryDoubleOperations.benchmark  20x200,20x20x200     JVM       DIVIDE  thrpt   15  177.003 ± 1.889  ops/s
BroadcastBinaryDoubleOperations.benchmark  20x200,20x20x200    ND4J        TIMES  thrpt   15  221.029 ± 2.052  ops/s
BroadcastBinaryDoubleOperations.benchmark  20x200,20x20x200    ND4J       DIVIDE  thrpt   15  203.764 ± 1.351  ops/s
```

After this PR:

```
Benchmark                                            (dims)  (impl)  (operation)   Mode  Cnt    Score   Error  Units
BroadcastBinaryDoubleOperations.benchmark  20x200,20x20x200     JVM        TIMES  thrpt   15  215.324 ± 1.241  ops/s
BroadcastBinaryDoubleOperations.benchmark  20x200,20x20x200     JVM       DIVIDE  thrpt   15  215.468 ± 1.997  ops/s
BroadcastBinaryDoubleOperations.benchmark  20x200,20x20x200    ND4J        TIMES  thrpt   15  213.266 ± 8.110  ops/s
BroadcastBinaryDoubleOperations.benchmark  20x200,20x20x200    ND4J       DIVIDE  thrpt   15  200.814 ± 7.031  ops/s
```